Fix some strict type violations reported by PhpStorm (strict types <3)

+++ src/pocketmine/entity/Projectile.php
@@ -74,7 +74,7 @@ abstract class Projectile extends Entity{
 	 * @return int
 	 */
 	public function getResultDamage() : int{
-		return ceil(sqrt($this->motionX ** 2 + $this->motionY ** 2 + $this->motionZ ** 2) * $this->damage);
+		return (int) ceil(sqrt($this->motionX ** 2 + $this->motionY ** 2 + $this->motionZ ** 2) * $this->damage);
 	}
 
 	public function onCollideWithEntity(Entity $entity){
